requirements.txt
einops~=0.8.0
kornia==0.2.0
numpy~=1.23.5
opencv_python==4.5.3.56
scikit_image==0.19.2
scikit_learn==1.1.3
scipy~=1.10.1
tensorboardX==2.5.1
timm==0.4.12
torch~=2.2.0+cu121

h5py~=3.11.0
opencv-python~=4.5.3.56
scikit-learn~=1.2.2
scikit-image~=0.19.2
tqdm~=4.66.4
pillow~=10.4.0
matplotlib~=3.7.0
utils~=1.0.2
transformers~=4.41.2
torchvision~=0.9.1
torchmetrics~=1.4.0.post0